* Fix for  [error: call of overloaded 'toString(size_t&)' is ambiguous] on os x
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@830 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/fileutils.cpp
@@ -340,7 +340,8 @@ const std::string getFilename(const std::string& pathname)
 		else
 			return "";
 	}
-}
+}
+
 
 const std::vector<std::string> getFilesInDir(const std::string& dir_path)
 {
@@ -970,7 +971,7 @@ void doUnitTests()
 				found_a = true;
 			if(files[i] == "b")
 				found_b = true;
-			conPrint("file[" + ::toString(i) + "]: '" + files[i] + "'");
+			conPrint("file[" + ::toString((int) i) + "]: '" + files[i] + "'");
 		}
 		testAssert(found_a);
 		testAssert(found_b);
